+++ packages/server/src/sdk/workspace/ai/rag/externalSources/sharepoint/sharepoint.spec.ts
